# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EduVijna Paper Evaluation API
  version: 0.1.0
paths:
...
  /api/v1/submissions/{id}/transcription/prepare:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      operationId: prepareTranscription
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Transcription prepare / retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranscriptionPrepareResult" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/submissions/{id}/transcription:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      operationId: getTranscriptionWorkspace
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Transcription review workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranscriptionWorkspace" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/submissions/{id}/transcription/finalize:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      operationId: finalizeTranscription
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Transcription READY — evaluation not enqueued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranscriptionFinalizeResult" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/answer-regions/{id}/transcription:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    put:
      operationId: putAnswerRegionTranscription
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ManualTranscriptionInput" }
      responses:
        "200":
          description: Human transcription version created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnswerRegionTranscription" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/answer-region-transcriptions/{id}/confirm:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      operationId: confirmAnswerRegionTranscription
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Transcription confirmed without rewriting AI confidence
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnswerRegionTranscription" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/answer-regions/{id}/crop:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      operationId: getAnswerRegionCrop
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Derived PNG crop bytes
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: Resource not found, including cross-tenant access
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
...
    ErrorEnvelope:
      $ref: "./schemas/error.schema.json"
...
    TranscriptionSegment:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, maxLength: 2000 }
        start: { type: [number, "null"] }
        end: { type: [number, "null"] }
    AnswerRegionTranscription:
      type: object
      required:
        [
          id,
          answer_region_id,
          version_number,
          source_type,
          status,
          unreadable,
          visual_only,
        ]
      properties:
        id: { type: string, format: uuid }
        answer_region_id: { type: string, format: uuid }
        version_number: { type: integer }
        source_type: { type: string, enum: [AI, HUMAN] }
        text: { type: [string, "null"] }
        latex: { type: [string, "null"] }
        segments:
          type: array
          items: { $ref: "#/components/schemas/TranscriptionSegment" }
        transcription_confidence:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        unreadable: { type: boolean }
        visual_only: { type: boolean }
        status:
          type: string
          enum: [PROPOSED, REVIEW_REQUIRED, CONFIRMED, SUPERSEDED]
        confirmed_by: { type: [string, "null"], format: uuid }
        confirmed_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ManualTranscriptionInput:
      type: object
      additionalProperties: false
      properties:
        text: { type: [string, "null"], maxLength: 20000 }
        latex: { type: [string, "null"], maxLength: 20000 }
        unreadable: { type: boolean }
        visual_only: { type: boolean }
        outcome: { type: string, enum: [TRANSCRIBED, UNREADABLE, VISUAL_ONLY] }
    TranscriptionProgress:
      type: object
      required: [reviewed, required, label]
      properties:
        reviewed: { type: integer }
        required: { type: integer }
        label: { type: string }
    TranscriptionWorkspaceItem:
      type: object
      required: [question_version_id, disposition, regions, requires_transcription]
      properties:
        question_version_id: { type: string, format: uuid }
        question_label: { type: string }
        disposition: { type: string, enum: [ANSWERED, BLANK] }
        mapping_state: { type: string }
        requires_transcription: { type: boolean }
        regions:
          type: array
          items:
            type: object
            required: [id, label, region_type, source_type]
            properties:
              id: { type: string, format: uuid }
              label: { type: string }
              region_type: { type: string }
              source_type: { type: string, enum: [AI, HUMAN] }
              detection_confidence: { type: number }
              crop_url: { type: [string, "null"] }
              page_image_url: { type: [string, "null"] }
              latest_ai_proposal:
                oneOf:
                  - { $ref: "#/components/schemas/AnswerRegionTranscription" }
                  - { type: "null" }
              active_transcription:
                oneOf:
                  - { $ref: "#/components/schemas/AnswerRegionTranscription" }
                  - { type: "null" }
              requires_transcription: { type: boolean }
    TranscriptionWorkspace:
      type: object
      required:
        [
          submission_id,
          workflow_state,
          transcription_state,
          automated_transcription_active,
          progress,
          items,
        ]
      properties:
        submission_id: { type: string, format: uuid }
        workflow_state: { type: string }
        transcription_state:
          type: string
          enum:
            [
              NOT_STARTED,
              QUEUED,
              RUNNING,
              REVIEW_REQUIRED,
              READY,
              FAILED,
              UNAVAILABLE,
            ]
        automated_transcription_active: { type: boolean }
        progress: { $ref: "#/components/schemas/TranscriptionProgress" }
        items:
          type: array
          items: { $ref: "#/components/schemas/TranscriptionWorkspaceItem" }
    TranscriptionPrepareResult:
      type: object
      required: [submission_id, workflow_state, transcription_state]
      properties:
        submission_id: { type: string, format: uuid }
        workflow_state: { type: string }
        transcription_state: { type: string }
        job_id: { type: [string, "null"], format: uuid }
    TranscriptionFinalizeResult:
      type: object
      required: [submission_id, workflow_state, transcription_state, evaluation_enqueued]
      properties:
        submission_id: { type: string, format: uuid }
        workflow_state: { type: string }
        transcription_state: { type: string, enum: [READY] }
        evaluation_enqueued: { type: boolean, enum: [false] }
...